Extract the authoring package (suggest/discover DOM scans)
Factor the bespoke browser-side JavaScript out of suggest and discover
into a shared `authoring` package so it lives in one place and can't drift:

- ScanCandidates(ctx, conn, max) — the data-testid/data-component selector
  candidate scan (behind `suggest`), returning ranked Candidate values.
- ScanLinks(ctx, conn) — the same-host a[href] pathname scan (behind
  `discover`).
- NormalizePath — folds a concrete path into a route pattern by replacing
  dynamic (numeric/UUID/slug) segments with "*".

The commands keep their flag parsing, corpus loading, filtering, grouping,
and printing — only the DOM JS and URL-pattern logic move. `gap` is
untouched: it uses no bespoke JS (extract + match + coverage T3).

Also dedup discover's hand-rolled routePatternToRegex — a copy of
sightmap/corpus.go's route matcher — by exporting sightmap.MatchRoute and
using it for survey.yaml glob matching.

Verified against the burrito example: suggest and discover run cleanly
(the app has no data-* attributes or <a href> links, so both correctly
report nothing); full suite green, plus a new NormalizePath test.
